requirements.txt
beautifulsoup4
requests
python-dotenv
newspaper3k
lxml_html_clean
nltk
